requirements.txt
flask
google-cloud-aiplatform
pillow>=10.4.0